후기 수정 및 삭제 구현 && 내가 남긴 후기 조회 구현

+++ V2.0/PDFLO_V2.0/src/main/java/travelbeeee/PDFLO_V20/controller/CommentController.java
@@ -4,9 +4,13 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import travelbeeee.PDFLO_V20.domain.dto.CommentModifyDto;
+import travelbeeee.PDFLO_V20.domain.entity.Comment;
 import travelbeeee.PDFLO_V20.domain.form.CommentForm;
 import travelbeeee.PDFLO_V20.exception.PDFLOException;
 import travelbeeee.PDFLO_V20.service.CommentService;
@@ -16,6 +20,8 @@ import javax.naming.Binding;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Controller
 @RequiredArgsConstructor
@@ -43,10 +49,59 @@ public class CommentController {
     @PostMapping("/comment/delete/{itemId}/{commentId}")
     public String deleteComment(HttpSession httpSession, @PathVariable("commentId") Long commentId,
                                 @PathVariable("itemId") Long itemId) throws PDFLOException {
+        PermissionChecker.checkPermission(httpSession);
+
         Long memberId = (Long) httpSession.getAttribute("id");
 
         commentService.deleteComment(memberId, commentId);
 
         return "redirect:/item/" + itemId;
+    }
+
+    @GetMapping("/comment/modify/{itemId}/{commentId}")
+    public String modifyCommentForm(HttpSession httpSession, @PathVariable("commentId") Long commentId,
+                                    @PathVariable("itemId") Long itemId,  Model model) throws PDFLOException {
+        PermissionChecker.checkPermission(httpSession);
+
+        Long memberId = (Long) httpSession.getAttribute("id");
+        Comment findComment = commentService.findByIdAndMember(commentId, memberId);
+
+        CommentModifyDto comment = new CommentModifyDto(findComment);
+
+        model.addAttribute("commentForm", new CommentForm());
+        model.addAttribute("comment", comment);
+        model.addAttribute("itemId", itemId);
+
+        return "/comment/modifyForm";
+    }
+
+    @PostMapping("/comment/modify/{itemId}/{commentId}")
+    public String modifyComment(HttpSession httpSession, @Valid CommentForm commentForm, BindingResult bindingResult,
+                                @PathVariable("itemId") Long itemId, @PathVariable("commentId") Long commentId) throws PDFLOException {
+        PermissionChecker.checkPermission(httpSession);
+
+        if (bindingResult.hasErrors()) {
+            return "redirect:/comment/modify/" + itemId + "/" + commentId;
+        }
+
+        Long memberId = (Long) httpSession.getAttribute("id");
+        commentService.modifyComment(memberId, commentId, commentForm);
+
+        return "redirect:/item/" + itemId;
+    }
+
+    @GetMapping("/comment/member")
+    public String myComment(HttpSession httpSession, Model model) throws PDFLOException {
+        PermissionChecker.checkPermission(httpSession);
+
+        Long memberId = (Long) httpSession.getAttribute("id");
+        List<Comment> comments = commentService.findAllWithItemByMember(memberId);
+        List<CommentItemDto> commentList = comments.stream()
+                .map(c -> new CommentItemDto(c))
+                .collect(Collectors.toList());
+
+        model.addAttribute("commentList", commentList);
+
+        return "/member/mycomment";
     }
 }
